+++ src/assets/styles/GlobalStyles.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const StyledApp = styled.div`
+  color: white;
+  min-height: 100vh;
+  max-width: 100vw;
+`;
+
+export const AlbumContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  margin-left: 20px;
+  margin-right: 20px;
+`;
