+++ src/pages/Album.js
@@ -5,20 +5,19 @@ import axios from 'axios';
 import Sheets from '../components/Sheets';
 import auth from '../auth';
 
-import styled from 'styled-components';
-
-export const AlbumContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  text-align: left;
-  margin-left: 20px;
-`;
+import { AlbumContainer } from '../assets/styles/GlobalStyles';
 
 function Album() {
   const [albumData, setAlbumData] = useState({
     films: auth.films,
     people: auth.people,
     starships: auth.starships,
+  });
+
+  const [toggle, setToggle] = useState({
+    films: false,
+    people: false,
+    starships: false,
   });
 
   const handleClick = async (name) => {
@@ -63,22 +62,56 @@ function Album() {
     }
   };
 
+  const handleToggle = (name) => {
+    console.log(name);
+    if (name === 'Películas') {
+      if (toggle.films === true) {
+        setToggle({ ...toggle, films: false });
+      }
+      if (toggle.films === false) {
+        setToggle({ ...toggle, films: true });
+      }
+    }
+    if (name === 'Personajes') {
+      if (toggle.people === true) {
+        setToggle({ ...toggle, people: false });
+      }
+      if (toggle.people === false) {
+        setToggle({ ...toggle, people: true });
+      }
+    }
+    if (name === 'Naves') {
+      if (toggle.starships === true) {
+        setToggle({ ...toggle, starships: false });
+      }
+      if (toggle.starships === false) {
+        setToggle({ ...toggle, starships: true });
+      }
+    }
+  };
+
   return (
     <AlbumContainer>
       <Sheets
         name="Películas"
         sheetType={albumData.films}
         handleClick={handleClick}
+        handleToggle={handleToggle}
+        toggle={toggle.films}
       />
       <Sheets
         name="Personajes"
         sheetType={albumData.people}
         handleClick={handleClick}
+        handleToggle={handleToggle}
+        toggle={toggle.people}
       />
       <Sheets
         name="Naves"
         sheetType={albumData.starships}
         handleClick={handleClick}
+        handleToggle={handleToggle}
+        toggle={toggle.starships}
       />
     </AlbumContainer>
   );
