+++ src/parts/Header.js
@@ -1,20 +1,8 @@
-import styled from 'styled-components';
-
-export const HeaderContainer = styled.header`
-  background-color: #000;
-  height: 100px;
-  opacity: 0.7;
-  border-bottom: 2px dashed #75C89C;
-`;
-
-export const LogoContainer = styled.figure`
-  height: 90%;
-`;
-
-export const StarWarsLogo = styled.img`
-  height: 100%;
-  filter: invert(1);
-`;
+import {
+  HeaderContainer,
+  LogoContainer,
+  StarWarsLogo,
+} from '../assets/styles/HeaderStyles';
 
 function Header() {
   return (
